feat(backend): add storing listing translations
